+++ Sources/TuistLoader/Models+ManifestMappers/Project+ManifestMapper.swift
@@ -12,6 +12,7 @@ extension TuistCore.Project {
     static func from(manifest: ProjectDescription.Project,
                      generatorPaths: GeneratorPaths) throws -> TuistCore.Project {
         let name = manifest.name
+        let organizationName = manifest.organizationName
         let settings = try manifest.settings.map { try TuistCore.Settings.from(manifest: $0, generatorPaths: generatorPaths) }
         let targets = try manifest.targets.map { try TuistCore.Target.from(manifest: $0, generatorPaths: generatorPaths) }
         let schemes = try manifest.schemes.map { try TuistCore.Scheme.from(manifest: $0, generatorPaths: generatorPaths) }
@@ -19,6 +20,7 @@ extension TuistCore.Project {
         let packages = try manifest.packages.map { try TuistCore.Package.from(manifest: $0, generatorPaths: generatorPaths) }
         return Project(path: generatorPaths.manifestDirectory,
                        name: name,
+                       organizationName: organizationName,
                        settings: settings ?? .default,
                        filesGroup: .group(name: "Project"),
                        targets: targets,
@@ -30,6 +32,7 @@ extension TuistCore.Project {
     func adding(target: TuistCore.Target) -> TuistCore.Project {
         Project(path: path,
                 name: name,
+                organizationName: organizationName,
                 fileName: fileName,
                 settings: settings,
                 filesGroup: filesGroup,
@@ -42,6 +45,20 @@ extension TuistCore.Project {
     func replacing(fileName: String?) -> TuistCore.Project {
         Project(path: path,
                 name: name,
+                organizationName: organizationName,
+                fileName: fileName,
+                settings: settings,
+                filesGroup: filesGroup,
+                targets: targets,
+                packages: packages,
+                schemes: schemes,
+                additionalFiles: additionalFiles)
+    }
+
+    func replacing(organizationName: String?) -> TuistCore.Project {
+        Project(path: path,
+                name: name,
+                organizationName: organizationName,
                 fileName: fileName,
                 settings: settings,
                 filesGroup: filesGroup,