+++ fixtures/app_with_organization_name_project/Project.swift
@@ -1,0 +1,12 @@
+import ProjectDescription
+
+let project = Project(name: "App",
+                      organizationName: "Tuist",
+                      targets: [
+                          Target(name: "App",
+                                 platform: .iOS,
+                                 product: .app,
+                                 bundleId: "io.tuist.app",
+                                 infoPlist: "Info.plist",
+                                 sources: "App/**")
+])